Improved code structuring
