+++ app/containers/styles/AppStyle.js
@@ -1,0 +1,20 @@
+
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: '#ffffff',
+  },
+  subcontainer: {
+    flex: 1,
+    justifyContent: 'center',
+    alignItems: 'center'
+  },
+  textStyle: {
+    fontSize: 20,
+    fontWeight: '500'
+  }
+});
+
+export  default styles;
