+++ package.json
@@ -7,7 +7,7 @@
     "test": "jest",
     "clean": "rm -rf $TMPDIR/react-* && watchman watch-del-all && npm cache clean --force",
     "clean:android": "cd android/ && ./gradlew clean && cd .. && react-native run-android",
-    "newclear": "rm -rf $TMPDIR/react-* && watchman watch-del-all && rm -rf ios/build && rm -rf node_modules/ && npm cache clean --force && npm i",
+    "cleanAll": "rm -rf $TMPDIR/react-* && watchman watch-del-all && rm -rf ios/build && rm -rf node_modules/ && npm cache clean --force && npm i",
     "android:build": "cd android && ./gradlew assembleRelease && open './app/build/outputs/apk/'",
     "android:install": "cd android && ./gradlew assembleRelease && ./gradlew installRelease",
     "android:devices": "$ANDROID_HOME/platform-tools/adb devices",
